requirements.txt
fastapi
uvicorn
python-dotenv
telethon
